qt4: fix preferences widget compile

+++ modules/gui/qt4/components/preferences_widgets.cpp
@@ -541,7 +541,7 @@ void setfillVLCConfigCombo( const char *configname, intf_thread_t *p_intf,
        if( string )
            def = QVariant( qfu(p_config->value.psz) );
        else
-           def = QVariant( p_config->value.i );
+           def = QVariant( qlonglong( p_config->value.i ) );
 
        if(p_config->pf_update_list)
         {
